fix: determinism fixes for P2P lockstep — float→double, qsort stability, zeroed allocs, deterministic saves
Apply all actionable determinism fixes identified in the audit (issue #4):

MEDIUM severity:
- M1: float→double in city_migration_score() and struct city field (cityturn.c, city.h)
- M2: float→double in city illness calculations (city.c)
- M3: (int)→lround() for combat defender selection (combat.c)
- M4: Fix double→int truncation in CM qsort comparator (cm.c)
- M6: float→double in AI evaluations (daimilitary.c, daidiplomat.c,
      daidiplomacy.c, daidomestic.c)
- M7: Sort nation_hash by nation_number() before serialization (savegame3.c)

LOW severity:
- L1: Add player_number() tiebreaker to report qsort (report.c)
- L5: fc_malloc→fc_calloc at high-risk struct allocations (voting.c,
      unittools.c, unithand.c)
- L6: Remove wall-clock timer_read_seconds() from save files (savegame3.c)

Documentation:
- Update DETERMINISM_AUDIT.md with fix status for all 16 issues

Closes #15, closes #41

+++ ai/default/daidiplomat.c
@@ -565,7 +565,7 @@ static struct city *dai_diplomat_defend(struct ai_type *ait,
     /* This formula may not be optimal, but it works. */
     if (move_cost > best_dist) {
       /* punish city for being so far away */
-      urgency /= (float) (move_cost / best_dist);
+      urgency /= (double) (move_cost / best_dist);
     }
 
     if (urgency > best_urgency) {
